+++ src/helpers/Field.php
@@ -18,7 +18,7 @@ use Craft;
 use craft\base\Element;
 use craft\base\Field as BaseField;
 use craft\ckeditor\Field as CKEditorField;
-use craft\elements\MatrixBlock;
+use craft\elements\Entry;
 use craft\elements\User;
 use craft\fields\Assets as AssetsField;
 use craft\fields\Matrix as MatrixField;
@@ -32,8 +32,6 @@ use nystudio107\seomatic\fields\SeoSettings as SeoSettingsField;
 use nystudio107\seomatic\Seomatic;
 use nystudio107\seomatic\services\MetaBundles;
 use verbb\doxter\fields\Doxter as DoxterField;
-use verbb\supertable\elements\SuperTableBlockElement as SuperTableBlock;
-use verbb\supertable\fields\SuperTableField;
 use yii\base\InvalidConfigException;
 
 /**
@@ -60,7 +58,6 @@ class Field
             RedactorField::class,
             TagsField::class,
             NeoField::class,
-            SuperTableField::class,
             PreparseFieldType::class,
             DoxterField::class,
         ],
@@ -70,7 +67,6 @@ class Field
         self::BLOCK_FIELD_CLASS_KEY => [
             MatrixField::class,
             NeoField::class,
-            SuperTableField::class,
         ],
         self::SEO_SETTINGS_CLASS_KEY => [
             SeoSettingsField::class,
@@ -98,11 +94,6 @@ class Field
      */
     public static $neoFieldsOfTypeCache = [];
 
-    /**
-     * @var array Memoization cache
-     */
-    public static $superTableFieldsOfTypeCache = [];
-
     // Static Methods
     // =========================================================================
 
@@ -120,7 +111,8 @@ class Field
         string      $fieldClassKey,
         FieldLayout $layout,
         bool        $keysOnly = true,
-    ): array {
+    ): array
+    {
         $foundFields = [];
         if (!empty(self::FIELD_CLASSES[$fieldClassKey])) {
             // Cache me if you can
@@ -163,7 +155,8 @@ class Field
         Element $element,
         string  $fieldClassKey,
         bool    $keysOnly = true,
-    ): array {
+    ): array
+    {
         $foundFields = [];
         $layout = $element->getFieldLayout();
         if ($layout !== null) {
@@ -274,7 +267,8 @@ class Field
         string $sourceHandle,
         string $fieldClassKey,
         bool   $keysOnly = true,
-    ): array {
+    ): array
+    {
         $foundFields = [];
         $layouts = [];
         // Get the layouts
@@ -297,30 +291,30 @@ class Field
     }
 
     /**
-     * Return all of the fields in the $matrixBlock of the type $fieldType class
+     * Return all of the fields in the $matrixEntry of the type $fieldType class
      *
-     * @param MatrixBlock $matrixBlock
+     * @param Entry $matrixEntry
      * @param string $fieldType
      * @param bool $keysOnly
      *
      * @return array
      */
-    public static function matrixFieldsOfType(MatrixBlock $matrixBlock, string $fieldType, bool $keysOnly = true): array
+    public static function matrixFieldsOfType(Entry $matrixEntry, string $fieldType, bool $keysOnly = true): array
     {
         $foundFields = [];
 
         try {
-            $matrixBlockTypeModel = $matrixBlock->getType();
+            $matrixEntryTypeModel = $matrixEntry->getType();
         } catch (InvalidConfigException $e) {
-            $matrixBlockTypeModel = null;
+            $matrixEntryTypeModel = null;
         }
-        if ($matrixBlockTypeModel) {
+        if ($matrixEntryTypeModel) {
             // Cache me if you can
-            $memoKey = $fieldType . $matrixBlock->id . ($keysOnly ? 'keys' : 'nokeys');
+            $memoKey = $fieldType . $matrixEntry->id . ($keysOnly ? 'keys' : 'nokeys');
             if (!empty(self::$matrixFieldsOfTypeCache[$memoKey])) {
                 return self::$matrixFieldsOfTypeCache[$memoKey];
             }
-            $fields = $matrixBlockTypeModel->getCustomFields();
+            $fields = $matrixEntryTypeModel->getCustomFields();
             /** @var BaseField $field */
             foreach ($fields as $field) {
                 if ($field instanceof $fieldType) {
@@ -372,52 +366,6 @@ class Field
             }
             // Cache for future use
             self::$neoFieldsOfTypeCache[$memoKey] = $foundFields;
-        }
-
-        return $foundFields;
-    }
-
-    /**
-     * Return all of the fields in the $superTableBlock of the type $fieldType class
-     *
-     * @param SuperTableBlock $superTableBlock
-     * @param string $fieldType
-     * @param bool $keysOnly
-     *
-     * @return array
-     */
-    public static function superTableFieldsOfType(SuperTableBlock $superTableBlock, string $fieldType, bool $keysOnly = true): array
-    {
-        $foundFields = [];
-
-        try {
-            $superTableBlockTypeModel = $superTableBlock->getType();
-        } catch (InvalidConfigException $e) {
-            $superTableBlockTypeModel = null;
-        }
-        if ($superTableBlockTypeModel) {
-            // Cache me if you can
-            $memoKey = $fieldType . $superTableBlock->id . ($keysOnly ? 'keys' : 'nokeys');
-            if (!empty(self::$superTableFieldsOfTypeCache[$memoKey])) {
-                return self::$superTableFieldsOfTypeCache[$memoKey];
-            }
-            /** @var ?FieldLayout $layout */
-            // The SuperTableBlockType class lacks @mixin FieldLayoutBehavior in its annotations
-            /** @phpstan-ignore-next-line */
-            $layout = $superTableBlockTypeModel->getFieldLayout();
-            $fieldElements = $layout->getCustomFieldElements();
-            foreach ($fieldElements as $fieldElement) {
-                $field = $fieldElement->getField();
-                if ($field instanceof $fieldType) {
-                    $foundFields[$field->handle] = $field->name;
-                }
-            }
-            // Return only the keys if asked
-            if ($keysOnly) {
-                $foundFields = array_keys($foundFields);
-            }
-            // Cache for future use
-            self::$superTableFieldsOfTypeCache[$memoKey] = $foundFields;
         }
 
         return $foundFields;