+++ src/helpers/Sitemap.php
@@ -9,7 +9,7 @@ use craft\base\Event;
 use craft\console\Application as ConsoleApplication;
 use craft\db\Paginator;
 use craft\elements\Asset;
-use craft\elements\MatrixBlock;
+use craft\elements\Entry;
 use craft\errors\SiteNotFoundException;
 use craft\fields\Assets as AssetsField;
 use craft\queue\Queue;
@@ -24,7 +24,6 @@ use nystudio107\seomatic\models\MetaBundle;
 use nystudio107\seomatic\models\SitemapTemplate;
 use nystudio107\seomatic\Seomatic;
 use Throwable;
-use verbb\supertable\elements\SuperTableBlockElement as SuperTableBlock;
 use yii\base\Exception;
 use yii\caching\TagDependency;
 use yii\helpers\Html;
@@ -324,17 +323,14 @@ class Sitemap
                         );
                         foreach ($blockFields as $blockField) {
                             $blocks = $element[$blockField]->all();
-                            /** @var MatrixBlock[]|NeoBlock[]|SuperTableBlock[]|object[] $blocks */
+                            /** @var Entry[]|NeoBlock[]|object[] $blocks */
                             foreach ($blocks as $block) {
                                 $assetFields = [];
-                                if ($block instanceof MatrixBlock) {
+                                if ($block instanceof Entry) {
                                     $assetFields = FieldHelper::matrixFieldsOfType($block, AssetsField::class);
                                 }
                                 if ($block instanceof NeoBlock) {
                                     $assetFields = FieldHelper::neoFieldsOfType($block, AssetsField::class);
-                                }
-                                if ($block instanceof SuperTableBlock) {
-                                    $assetFields = FieldHelper::superTableFieldsOfType($block, AssetsField::class);
                                 }
                                 foreach ($assetFields as $assetField) {
                                     foreach ($block[$assetField]->all() as $asset) {
@@ -368,14 +364,11 @@ class Sitemap
                     );
                     foreach ($blockFields as $blockField) {
                         $blocks = $element[$blockField]->all();
-                        /** @var MatrixBlock[]|NeoBlock[]|SuperTableBlock[]|object[] $blocks */
+                        /** @var Entry[]|NeoBlock[]|object[] $blocks */
                         foreach ($blocks as $block) {
                             $assetFields = [];
-                            if ($block instanceof MatrixBlock) {
+                            if ($block instanceof Entry) {
                                 $assetFields = FieldHelper::matrixFieldsOfType($block, AssetsField::class);
-                            }
-                            if ($block instanceof SuperTableBlock) {
-                                $assetFields = FieldHelper::superTableFieldsOfType($block, AssetsField::class);
                             }
                             if ($block instanceof NeoBlock) {
                                 $assetFields = FieldHelper::neoFieldsOfType($block, AssetsField::class);