+++ src/seoelements/SeoShopifyProduct.php
@@ -137,7 +137,7 @@ class SeoShopifyProduct implements SeoElementInterface
                 );
                 $html = '';
                 Seomatic::$view->registerAssetBundle(SeomaticAsset::class);
-                /** @var  $product Product */
+                /** @var Product $product */
                 $product = $event->sender ?? null;
                 if ($product !== null && $product->uri !== null) {
                     Seomatic::$plugin->metaContainers->previewMetaContainers($product->uri, $product->siteId, true);
@@ -186,7 +186,8 @@ class SeoShopifyProduct implements SeoElementInterface
         MetaBundle $metaBundle,
         int        $elementId,
         int        $siteId,
-    ) {
+    )
+    {
         return Product::find()
             ->id($elementId)
             ->siteId($siteId)
@@ -248,7 +249,7 @@ class SeoShopifyProduct implements SeoElementInterface
      *
      * @param int $sourceId
      *
-     * @return null
+     * @return ProductType|null
      */
     public static function sourceModelFromId(int $sourceId)
     {
@@ -260,7 +261,7 @@ class SeoShopifyProduct implements SeoElementInterface
      *
      * @param string $sourceHandle
      *
-     * @return null
+     * @return ProductType|null
      */
     public static function sourceModelFromHandle(string $sourceHandle)
     {